Fix a bug on database/population/country.php and add more testing

+++ tests/base.php
@@ -27,6 +27,9 @@
         public function setUp() {
             $tables = dbListTables();
             foreach ( $tables as $table ) {
+                if ( $table === 'countries' ) {
+                    continue;
+                }
                 db( 'TRUNCATE TABLE ' . $table );
             }
         }
